apps/data-engine/requirements.txt
fastapi
uvicorn
pandas
httpx
requests
playwright
beautifulsoup4
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings
sqlalchemy>=2.0
alembic
redis
apscheduler
pymysql
cryptography
elasticsearch
crawl4ai
